refactor: move expr into core.expr
