+++ mplang/core/expr/evaluator.py
@@ -24,9 +24,7 @@ from __future__ import annotations
 from typing import Any
 
 from mplang.core.comm import ICommunicator
-from mplang.core.mask import Mask
-from mplang.core.pfunc import PFunction, PFunctionHandler
-from mplang.expr.ast import (
+from mplang.core.expr.ast import (
     AccessExpr,
     CallExpr,
     CondExpr,
@@ -40,7 +38,9 @@ from mplang.expr.ast import (
     VariableExpr,
     WhileExpr,
 )
-from mplang.expr.visitor import ExprVisitor
+from mplang.core.expr.visitor import ExprVisitor
+from mplang.core.mask import Mask
+from mplang.core.pfunc import PFunction, PFunctionHandler
 
 
 class Evaluator(ExprVisitor):
