+++ mplang/core/interp.py
@@ -26,10 +26,10 @@ from abc import abstractmethod
 from collections.abc import Sequence
 from typing import Any, cast
 
+from mplang.core.expr.ast import Expr, VariableExpr
 from mplang.core.mpobject import MPContext, MPObject
 from mplang.core.mptype import MPType, TensorLike
 from mplang.core.tracer import TracedFunction
-from mplang.expr.ast import Expr, VariableExpr
 from mplang.utils.func_utils import var_demorph, var_morph
