+++ mplang/core/tracer.py
@@ -61,12 +61,12 @@ from dataclasses import dataclass
 from typing import Any, cast
 
 from mplang.core.context_mgr import with_ctx
+from mplang.core.expr.ast import Expr, FuncDefExpr, TupleExpr, VariableExpr
+from mplang.core.expr.printer import Printer
 from mplang.core.mask import Mask
 from mplang.core.mpobject import MPContext, MPObject
 from mplang.core.mptype import MPType
 from mplang.core.pfunc import get_fn_name
-from mplang.expr.ast import Expr, FuncDefExpr, TupleExpr, VariableExpr
-from mplang.expr.printer import Printer
 from mplang.utils.func_utils import MorphStruct, var_demorph, var_morph
