+++ tests/core/expr/test_utils.py
@@ -15,8 +15,13 @@
 import pytest
 
 from mplang.core.dtype import FLOAT32, INT32
+from mplang.core.expr import (
+    deduce_mask,
+    ensure_scalar,
+    ensure_tensorlist_equal,
+    type_equal,
+)
 from mplang.core.mask import Mask
-from mplang.expr import deduce_mask, ensure_scalar, ensure_tensorlist_equal, type_equal
 
 
 class MockTensor:
